created a user route file; added path param to question route
