+++ index.ts
@@ -1,13 +1,11 @@
-import express, { Request, Response } from "express";
-import { User } from "./src/database/entity/user.entity";
+import express, { Response } from "express";
 import dotenv from "dotenv";
 import myDataSource from "./app-data-source";
-import UserService from "./src/database/repositories/users.service";
-import UserPayload from "./src/models/User.model";
 import QuestionsRoute from "./src/routes/questions";
+import UserRoute from "./src/routes/users";
 
-const cors = require("cors");
-const bodyParser = require("body-parser");
+import cors from "cors";
+import bodyParser from "body-parser";
 
 myDataSource
   .initialize()
@@ -27,35 +25,13 @@ app.use(
 );
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
+
+// ROUTES
 app.use("/question", QuestionsRoute);
+app.use("/user", UserRoute);
 
 app.get("/", async (_, res: Response) => {
   res.send("Express + TypeScript Server");
-});
-
-app.get("/users", async (_, res: Response) => {
-  const users = await myDataSource
-    .getRepository(User)
-    .createQueryBuilder("users")
-    .getMany();
-  res.send(users);
-});
-
-app.post("/user", async (req: Request, res: Response) => {
-  const userService = new UserService(myDataSource);
-
-  try {
-    let reqBody = req.body as UserPayload;
-    const trimAuthId = reqBody.authId.split("|")[1];
-    reqBody = { ...reqBody, authId: trimAuthId };
-    const user = await userService.getUserByAuthId(trimAuthId);
-    if (!user) {
-      await userService.saveUser(reqBody);
-      res.send("User added to DB");
-    }
-  } catch (err) {
-    console.log("err", err);
-  }
 });
 
 app.listen(port, () => {
