+++ package.json
@@ -24,6 +24,7 @@
     "typeorm": "^0.3.15"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
     "@types/node": "^18.16.3",
     "typescript": "^5.0.4"
