+++ src/routes/questions.ts
@@ -3,7 +3,12 @@ import express, { Request, Response } from "express";
 const router = express.Router();
 
 router.get("/", (req: Request, res: Response) => {
-  res.send("hello user");
+  res.send("GET Question");
+});
+
+router.get("/:id", (req: Request, res: Response) => {
+  const questionId = req.params.id;
+  res.send(`Question ${questionId}`);
 });
 
 export default router;
